refactor arithmetic fucntions
- add unit testing
- refactor arithmetic fucntions to  better handle errors

+++ src/functions/v1/addition/handler.ts
@@ -12,7 +12,7 @@ const addition = async (event, context) => {
   let status = 503;
   try {
     const {numberA, numberB} = event.queryStringParameters;
-    internalResponse = additionService(parseInt(numberA), parseInt(numberB));
+    internalResponse = additionService(parseFloat(numberA), parseFloat(numberB));
     if (!internalResponse.error) {
       status = 200;
     }
